==> Rmarkdown/01b_BIMP_DataInput_Polyandry.rmd <==
---
title: "01b_BIMP_DataInput_Polyandry.rmd"
author: "Garett Slater"
date: "May 26 2022"
output:
  html_document: default
  pdf_document: default
---

```{r knitr setup, include=FALSE,  eval=TRUE, echo=FALSE, warning=FALSE}
library(knitr)
knitr::opts_chunk$set(eval=TRUE, cache=FALSE, message=FALSE, warning=FALSE, 
                      comment = "")
```

```{r, include=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Code")
source("Polyandry_Evo_Functions_V2.R", local = knitr::knit_global())
```


# Data Input
```{r data}
########################################
##                                    ##       
##        Resource Files              ##
##                                    ## 
########################################
library(pacman)
p_load(biomatRt,tximport,DESeq2,dplyr,plyranges,ggplot2,plyr,ggpubr,reshape2,UpSetR,tispec,tidyverse,rstatix,readxl)

#Function Classes
func<-read.csv("~/OneDrive/Projects/Experimental Protocols/Bioinformatics/Proteomics/drone_function_final.csv")

########################################
##                                    ##       
## 1. Gene Datasets and Figure 1A     ##
##                                    ## 
########################################
########################################
##               BIMP                 ##
########################################
#Description Files
BIMP_Sample = vector(mode = "list",length=3)
BIMP_Sample[[1]]$Sample<-read.csv("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources/BIMP_samp.csv")
BIMP_Sample[[2]]$ortho<-read.csv("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources/BIMP_ortho.csv")
BIMP_Sample[[3]]$tx2gene<-read.csv("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources/BIMP_tx2gene.csv")
```

# Gene Expression, Kallisto, and DESEq analysis
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Abundance")
#DESeq Analysis
#Run kallisto with Final (Tissue Type) as design and a count cutoff of 10
BIMP_final = vector(mode = "list", length = 4)

#Gondal
Sample<-BIMP_Sample[[1]]$Sample[which(BIMP_Sample[[1]]$Sample$Tissue_Type=="Gonadal"),]
dds<-kallisto(Sample,BIMP_Sample[[3]]$tx2gene,Type=~Final)
BIMP_final[[1]]$Gonadal<-contrast("Final","MG","FG","BIMP_")

BIMP_final[[1]]$Gonadal$Bias[BIMP_final[[1]]$Gonadal$MB == 1]<-"MG"
BIMP_final[[1]]$Gonadal$Bias[BIMP_final[[1]]$Gonadal$FB == 1]<-"FG"
BIMP_final[[1]]$Gonadal$species<-"BIMP"
```

# Gonadal Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
pdf(file = "1-BIMP-UpsetPlot-Gonadal.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(BIMP_final[[1]]$Gonadal[,2:3],order.by=c("freq"))
dev.off()
upset(BIMP_final[[1]]$Gonadal[,2:3],order.by=c("freq"))
```

# Gonadal Ortho Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
#Gondal-Ortho Only
BIMP_final[[3]]$Gonadal_Ortho = merge(BIMP_Sample[[2]]$ortho, BIMP_final[[1]]$Gonadal,by="BIMP_protein_id",all=TRUE)
BIMP_final[[3]]$Gonadal_Ortho<-BIMP_final[[3]]$Gonadal_Ortho %>%
                                      mutate(across(everything(), ~ ifelse(is.na(.), 0, .)))  %>%
                                      mutate(UB=ifelse(MB+FB==0,1,0))%>%
                                      filter(protein_id!=0) %>%
                                      select(protein_id,MB,FB,UB)

BIMP_final[[3]]$Gonadal_Ortho$Bias[BIMP_final[[3]]$Gonadal_Ortho$MB == 1]<-"MG"
BIMP_final[[3]]$Gonadal_Ortho$Bias[BIMP_final[[3]]$Gonadal_Ortho$FB == 1]<-"FG"
BIMP_final[[3]]$Gonadal_Ortho$Bias[BIMP_final[[3]]$Gonadal_Ortho$UB == 1]<-"UB"
BIMP_final[[3]]$Gonadal_Ortho$species<-"BIMP"

pdf(file = "1-BIMP-UpsetPlot-GonadalOrtho.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(BIMP_final[[3]]$Gonadal_Ortho[,2:4],order.by=c("freq"))
dev.off()
upset(BIMP_final[[3]]$Gonadal_Ortho[,2:4],order.by=c("freq"))
```

```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Abundance")
#Somatic - Kallisto function and then contrast function
Sample<-BIMP_Sample[[1]]$Sample[which(BIMP_Sample[[1]]$Sample$Tissue_Type=="Somatic"),]
dds<-kallisto(Sample,BIMP_Sample[[3]]$tx2gene,Type=~Final)
BIMP_final[[2]]$Somatic<-contrast("Final","MS","FS","BIMP_")

BIMP_final[[2]]$Somatic$Bias[BIMP_final[[2]]$Somatic$MB == 2]<-"MS"
BIMP_final[[2]]$Somatic$Bias[BIMP_final[[2]]$Somatic$FB == 2]<-"FS"
BIMP_final[[2]]$Somatic$species<-"BIMP"
```

# Somatic Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
pdf(file = "1-BIMP-UpsetPlot-Somatic.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(BIMP_final[[2]]$Somatic[,2:3],order.by=c("freq"))
dev.off()
upset(BIMP_final[[2]]$Somatic[,2:3],order.by=c("freq"))
```

# Somatic Ortho Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
#Somatic-Ortho Only
BIMP_final[[4]]$Somatic_Ortho = merge(BIMP_Sample[[2]]$ortho, BIMP_final[[2]]$Somatic,by="BIMP_protein_id",all=TRUE)
BIMP_final[[4]]$Somatic_Ortho<-BIMP_final[[4]]$Somatic_Ortho %>%
                                      mutate(across(everything(), ~ ifelse(is.na(.), 0, .)))  %>%
                                      mutate(UB=ifelse(MB+FB==0,1,0)) %>%
                                      filter(protein_id!=0) %>%
                                      select(protein_id,MB,FB,UB)

BIMP_final[[4]]$Somatic_Ortho$Bias[BIMP_final[[4]]$Somatic_Ortho$MB == 1]<-"MS"
BIMP_final[[4]]$Somatic_Ortho$Bias[BIMP_final[[4]]$Somatic_Ortho$FB == 1]<-"FS"
BIMP_final[[4]]$Somatic_Ortho$Bias[BIMP_final[[4]]$Somatic_Ortho$UB == 1]<-"UB"
BIMP_final[[4]]$Somatic_Ortho$species<-"BIMP"

pdf(file = "1-BIMP-UpsetPlot-SomaticOrtho.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(BIMP_final[[4]]$Somatic_Ortho[,2:4],order.by=c("freq"))
dev.off()
upset(BIMP_final[[4]]$Somatic_Ortho[,2:4],order.by=c("freq"))

setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources")
save(BIMP_final,BIMP_Sample,file = "BIMP.RData")
```
