==> Rmarkdown/01d_TCARB_DataInput_Polyandry.rmd <==
---
title: "01d_TCARB_DataInput_Polyandry.rmd"
author: "Garett Slater"
date: "May 26 2022"
output:
  html_document: default
  pdf_document: default
---

```{r knitr setup, include=FALSE,  eval=TRUE, echo=FALSE, warning=FALSE}
library(knitr)
knitr::opts_chunk$set(eval=TRUE, cache=FALSE, message=FALSE, warning=FALSE, 
                      comment = "")
```

```{r, include=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Code")
source("Polyandry_Evo_Functions_V2.R", local = knitr::knit_global())
```

# Data Input
```{r data}
########################################
##                                    ##       
##        Resource Files              ##
##                                    ## 
########################################
library(pacman)
p_load(biomatRt,tximport,DESeq2,dplyr,plyranges,ggplot2,plyr,ggpubr,reshape2,UpSetR,tispec,tidyverse,rstatix,readxl)

########################################
##                                    ##       
## 1. Gene Datasets and Figure 1A     ##
##                                    ## 
########################################
########################################
##               BIMP                 ##
########################################
#Description Files
TCARB_Sample = vector(mode = "list",length=3)
TCARB_Sample[[1]]$Sample<-read.csv("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources/TCARB_samp.csv")
TCARB_Sample[[2]]$ortho<-read.csv("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources/TCARB_ortho.csv")
TCARB_Sample[[3]]$tx2gene<-read.csv("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources/TCARB_tx2gene.csv")
```


# Gene Expression, Kallisto, and DESEq analysis
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Abundance")
#DESeq Analysis
#Run kallisto with Final (Tissue Type) as design and a count cutoff of 10
TCARB_final = vector(mode = "list", length = 4)

#Gondal
Sample<-TCARB_Sample[[1]]$Sample[which(TCARB_Sample[[1]]$Sample$Tissue_Type=="Gonadal"),]
dds<-kallisto(Sample,TCARB_Sample[[3]]$tx2gene,Type=~Final)
TCARB_final[[1]]$Gonadal<-contrast("Final","MG","FG","TCARB_")
```

# Gonadal Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
TCARB_final[[1]]$Gonadal$Bias[TCARB_final[[1]]$Gonadal$MB == 1]<-"MG"
TCARB_final[[1]]$Gonadal$Bias[TCARB_final[[1]]$Gonadal$FB == 1]<-"FG"
TCARB_final[[1]]$Gonadal$species<-"TCARB"

pdf(file = "1-TCARB-UpsetPlot-Gonadal.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(TCARB_final[[1]]$Gonadal[,2:3],order.by=c("freq"))
dev.off()
upset(TCARB_final[[1]]$Gonadal[,2:3],order.by=c("freq"))
```

# Gonadal Ortho Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
#Gondal-Ortho Only
TCARB_final[[3]]$Gonadal_Ortho = merge(TCARB_Sample[[2]]$ortho, TCARB_final[[1]]$Gonadal,by="TCARB_protein_id",all=TRUE)
TCARB_final[[3]]$Gonadal_Ortho<-TCARB_final[[3]]$Gonadal_Ortho %>%
  mutate(across(everything(), ~ ifelse(is.na(.), 0, .)))  %>%
  mutate(UB=ifelse(MB+FB==0,1,0))%>%
  filter(protein_id!=0) %>%
  select(protein_id,MB,FB,UB)

TCARB_final[[3]]$Gonadal_Ortho$Bias[TCARB_final[[3]]$Gonadal_Ortho$MB == 1]<-"MG"
TCARB_final[[3]]$Gonadal_Ortho$Bias[TCARB_final[[3]]$Gonadal_Ortho$FB == 1]<-"FG"
TCARB_final[[3]]$Gonadal_Ortho$Bias[TCARB_final[[3]]$Gonadal_Ortho$UB == 1]<-"UB"
TCARB_final[[3]]$Gonadal_Ortho$species<-"TCARB"

pdf(file = "1-TCARB-UpsetPlot-Gonadal.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(TCARB_final[[3]]$Gonadal_Ortho[,2:4],order.by=c("freq"))
dev.off()
upset(TCARB_final[[3]]$Gonadal_Ortho[,2:4],order.by=c("freq"))
```

```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Abundance")
#Somatic - Kallisto function and then contrast function
Sample<-TCARB_Sample[[1]]$Sample[which(TCARB_Sample[[1]]$Sample$Tissue_Type=="Somatic"),]
dds<-kallisto(Sample,TCARB_Sample[[3]]$tx2gene,Type=~Final)
TCARB_final[[2]]$Somatic<-contrast("Final","MS","FS","TCARB_")
```

# Somatic Ortho Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
TCARB_final[[2]]$Somatic$Bias[TCARB_final[[2]]$Somatic$MB == 2]<-"MS"
TCARB_final[[2]]$Somatic$Bias[TCARB_final[[2]]$Somatic$FB == 2]<-"FS"
TCARB_final[[2]]$Somatic$species<-"TCARB"

pdf(file = "1-TCARB-UpsetPlot-Somatic.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(TCARB_final[[2]]$Somatic[,2:3],order.by=c("freq"))
dev.off()
upset(TCARB_final[[2]]$Somatic[,2:3],order.by=c("freq"))
```

# Somatic Ortho Upset
```{r echo=FALSE}
setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Graphics/RMD_Graphics")
#Somatic-Ortho Only
TCARB_final[[4]]$Somatic_Ortho = merge(TCARB_Sample[[2]]$ortho, TCARB_final[[2]]$Somatic,by="TCARB_protein_id",all=TRUE)
TCARB_final[[4]]$Somatic_Ortho<-TCARB_final[[4]]$Somatic_Ortho %>%
  mutate(across(everything(), ~ ifelse(is.na(.), 0, .)))  %>%
  mutate(UB=ifelse(MB+FB==0,1,0)) %>%
  filter(protein_id!=0) %>%
  select(protein_id,MB,FB,UB)

TCARB_final[[4]]$Somatic_Ortho$Bias[TCARB_final[[4]]$Somatic_Ortho$MB == 1]<-"MS"
TCARB_final[[4]]$Somatic_Ortho$Bias[TCARB_final[[4]]$Somatic_Ortho$FB == 1]<-"FS"
TCARB_final[[4]]$Somatic_Ortho$Bias[TCARB_final[[4]]$Somatic_Ortho$UB == 1]<-"UB"
TCARB_final[[4]]$Somatic_Ortho$species<-"TCARB"


pdf(file = "1-TCARB-UpsetPlot-SomaticOrtho.pdf", width = 12, height = 9);
par(cex = 1.6);
par(mar = c(0,4,2,0))
upset(TCARB_final[[4]]$Somatic_Ortho[,2:4],order.by=c("freq"))
dev.off()
upset(TCARB_final[[4]]$Somatic_Ortho[,2:4],order.by=c("freq"))

setwd("~/OneDrive/Manuscripts/Polyandry_GeneEvolution/Resources")
save(TCARB_final,TCARB_Sample,file = "TCARB.RData")
```
